move Flap class from binary_sensor to sensor and add a sensore base class
